[1.1.6] Packaged properly;
Took 4 minutes
